Add script for copying and renaming binaries - Created `ci/copy-bins.sh` to automate binary extraction - Used Docker to copy binaries from `zaino-ci` image
This is a temporary workaround. Since some tests look for bins in PATH and some in /test_binaries/bins, we have to support both.
Issue is, test_binaries lives in the rootdir which is mounted, hence overwriten.
One solution could be an entrypoint script that copies them over... But we want to move away from that pattern anyways (the one of having bins in test_binaries)